still can not fix upgrade profile
