+++ travlr/src/api/auth.ts
@@ -106,62 +106,3 @@ export const signupUser = async (
 		}
 	}
 }
-
-// export const fetchUserProfile = async (
-// 	userId: string
-// ): Promise<ApiResponse<UserProfile>> => {
-// 	try {
-// 		const response = await fetch(`${API_URL}/users/profile/${userId}`)
-// 		const textResponse = await response.text()
-// 		console.log("Raw response text:", textResponse)
-
-// 		const result = JSON.parse(textResponse)
-
-// 		if (response.ok) {
-// 			return { success: true, user: result.profile }
-// 		} else {
-// 			return {
-// 				success: false,
-// 				error: result.error || "Failed to fetch profile",
-// 			}
-// 		}
-// 	} catch (error) {
-// 		return { success: false, error: (error as Error).message }
-// 	}
-// }
-
-// export const updateUserProfile = async (
-// 	userId: string,
-// 	profileData: Partial<UserProfile>
-// ): Promise<ApiResponse<UserProfile>> => {
-// 	try {
-// 		const response = await axios.put(
-// 			`${API_URL}/users/profileUpdate/${userId}`,
-// 			profileData
-// 		)
-// 		return {
-// 			success: true,
-// 			user: response.data,
-// 		}
-// 	} catch (error) {
-// 		if (axios.isAxiosError(error)) {
-// 			return {
-// 				success: false,
-// 				error:
-// 					error.response?.data?.error ||
-// 					"An unexpected error occurred at updateUserProfile",
-// 			}
-// 		} else if (error instanceof Error) {
-// 			return {
-// 				success: false,
-// 				error:
-// 					error.message || "An unexpected error occurred at updateUserProfile",
-// 			}
-// 		} else {
-// 			return {
-// 				success: false,
-// 				error: "An unexpected error occurred at updateUserProfile",
-// 			}
-// 		}
-// 	}
-// }
